+++ R/parallel.R
@@ -26,25 +26,18 @@
 #'     package such as \code{\link[parallel]{clusterApply}} or
 #'     \code{\link[parallel]{parLapply}}.
 #'
+#' @inheritParams launch_remote
 #' @param n integer number of nodes (launched on the local machine unless 'url'
 #'     is specified).
 #' @param url (specify for remote nodes), the character URL on the host for
 #'     remote nodes to dial into, including a port accepting incoming connections,
 #'     e.g. 'tcp://10.75.37.40:5555'. Specify a URL starting 'tls+tcp://' to use
 #'     secure TLS connections.
-#' @param ssh_nodes (if 'url' is specified, for launching remote nodes via SSH)
-#'     a character vector of hostnames or IP addresses of the remote machines on
-#'     which to launch nodes, e.g. \code{c('10.75.37.90', 'nodename')}.
-#' @param ssh_port [default 22] numeric port number on which to connect.
-#' @param ssh_timeout [default 5] maximum time allowed for connection setup in
-#'     seconds.
-#' @param ssh_tunnel [default FALSE] logical value whether to use SSH reverse
-#'     tunnelling. If TRUE, a tunnel is created between the same local and
-#'     remote port specified as part of 'url'. 'url' in this case must be either
-#'     'localhost' or '127.0.0.1'. This is as the host listens at 'url' on the
-#'     local machine, whilst the nodes dial into the same 'url' on the remote
-#'     machine, and the SSH tunnel connects both ends.
 #' @param ... additional arguments passed onto \code{\link{daemons}}.
+#' @param rscript (for remote only) name / path of the Rscript executable on
+#'     the remote machine. The default assumes 'Rscript' is on the executable
+#'     search path. Prepend the full path if necessary. If launching on Windows,
+#'     'Rscript' should be replaced with 'Rscript.exe'.
 #'
 #' @return For \strong{make_cluster}: An object of class 'miraiCluster' and
 #'     'cluster'. Each 'miraiCluster' has an automatically assigned ID and 'n'
@@ -57,14 +50,14 @@
 #'     However, '...' arguments are passed onto \code{\link{daemons}} for
 #'     additional customisation if desired.
 #'
-#'     For remote nodes, the 'ssh_' arguments are an optional convenience
-#'     feature. If used, the number of nodes is inferred from the length of the
-#'     character vector 'ssh_nodes' and 'n' is disregarded if supplied. For ease
-#'     of use, SSH tunnelling assumes the same ports are available for forwarding
-#'     on both host and nodes, whilst this is not strictly necessary.
+#'     For remote nodes, the arguments 'command', 'args' and 'rscript' map
+#'     directly to the relevant arguments of \code{\link{launch_remote}}. If
+#'     launching remote nodes, the number of nodes is inferred from the length
+#'     of 'args' and 'n' is disregarded if supplied.
 #'
-#'     Alternatively, by specifying 'url' and 'n', nodes may also be launched by
-#'     other more customisable means, for example using \code{\link{launch_remote}}.
+#'     Alternatively, by specifying only 'url' and 'n', the host connection
+#'     listening at 'url' is set up, and nodes may be launched by other custom
+#'     means.
 #'
 #' @note Requires R >= 4.4 (currently R-devel). Clusters created with this
 #'     function will not work with prior R versions.
@@ -82,8 +75,10 @@
 #'   make_cluster(
 #'     # host URL accessible to nodes
 #'     url = "tcp://10.75.37.40:5555",
+#'     # launch nodes using SSH
+#'     command = "ssh",
 #'     # node IP / hostnames to connect to with custom port number
-#'     ssh = list(nodes = c("10.75.37.90", "10.75.37.91"), port = 222)
+#'     args = ssh_args(names = c("10.75.37.90", "10.75.37.91"), port = 222)
 #'   )
 #' , error = identity)
 #' if (inherits(cl, "cluster")) stop_cluster(cl)
@@ -92,33 +87,18 @@
 #'
 #' @export
 #'
-make_cluster <- function(n, url = NULL, ssh_nodes = character(), ssh_port = 22,
-                         ssh_timeout = 5, ssh_tunnel = FALSE, ...) {
+make_cluster <- function(n, url = NULL, ..., command = NULL, args = c("", "."), rscript = "Rscript") {
 
   id <- sprintf("`%d`", length(..))
 
   if (is.character(url)) {
 
-    if (length(ssh_nodes)) {
+    length(url) == 1L || stop(.messages[["single_url"]])
+
+    if (length(command)) {
       daemons(url = url, dispatcher = FALSE, resilience = FALSE, cleanup = 0L, ..., .compute = id)
-      for (node in ssh_nodes)
-        launch_remote(
-          url = 1L,
-          command = "ssh",
-          args = c(
-            if (ssh_tunnel) {
-              purl <- parse_url(url)
-              purl[["hostname"]] %in% c("localhost", "127.0.0.1") || stop(.messages[["tunnel_requires"]])
-              sprintf("-R %s:%s", purl[["port"]], purl[["host"]])
-            },
-            sprintf("-o ConnectTimeout=%s -fTp", as.character(ssh_timeout)),
-            as.character(ssh_port),
-            node,
-            .
-          ),
-          .compute = id
-        )
-      n <- length(ssh_nodes)
+      launch_remote(url = url, .compute = id, command = command, args = args, rscript = rscript)
+      n <- if (is.list(args)) length(args) else 1L
 
     } else {
       missing(n) && stop(.messages[["requires_n"]])
@@ -225,9 +205,5 @@ print.miraiNode <- function(x, ...) {
   invisible(x)
 
 }
-
-# internals --------------------------------------------------------------------
-
-node_unresolved <- function(node) unresolved(.subset2(node, "mirai"))
 
 # nocov end